+++ src/components/account/LootTracker.js
@@ -18,7 +18,7 @@ const buildLootRecord = record => (
 
 class LootTracker extends Component {
   componentDidMount() {
-    this.props.getReleases().then(() => this.props.fetchLoot())
+    this.props.fetchReleases().then(() => this.props.fetchLoot())
   }
 
   render({ loot }) {
