+++ src/components/app.js
@@ -12,8 +12,8 @@ import Redirect from './redirect'
 import Loader from './loader'
 import Async from './async'
 import { getStargazers } from '../modules/git'
-import { isLoggedIn, login } from '../modules/session'
 import NotFound from './not-found'
+import { isLoggedIn, login } from '../modules/account'
 
 const App = ({
   loading,
@@ -78,7 +78,7 @@ export default connect(
     stars: getStargazers(state),
     loggedIn: isLoggedIn(state),
     ...state.app,
-    ...state.session
+    ...state.account
   }),
   dispatch => bindActionCreators({ login }, dispatch)
 )(App)
