+++ src/index.js
@@ -2,7 +2,7 @@ import { h, render } from 'preact'
 import { Provider } from 'preact-redux'
 import App from './components/app'
 import configureStore from './store'
-import { sessionCheck } from './modules/session'
+import { sessionCheck } from './modules/account'
 import { fetchReleases } from './modules/git'
 
 // Create redux store
