+++ src/modules/index.js
@@ -1,3 +1,4 @@
+import accountReducer from './account'
 import appReducer from './app'
 import configReducer from './config'
 import geReducer from './ge'
@@ -5,12 +6,12 @@ import gitReducer from './git'
 import itemReducer from './item'
 import lootReducer from './loot'
 import runeliteReducer from './runelite'
-import sessionReducer from './session'
 import tagReducer from './tag'
 import xpReducer from './xp'
 
 // Combine all redux reducers into one root reducer
 export default {
+  account: accountReducer,
   app: appReducer,
   config: configReducer,
   ge: geReducer,
@@ -18,7 +19,6 @@ export default {
   item: itemReducer,
   loot: lootReducer,
   runelite: runeliteReducer,
-  session: sessionReducer,
   tag: tagReducer,
   xp: xpReducer
 }
